+++ Func.hpp
@@ -4,7 +4,7 @@
 namespace func
 {
     // lay khoang cach
-    int dist( int x, int y, int u, int v );
+    float dist( int x, int y, int u, int v );
     // va cham
     bool checkRect( SDL_Rect rect1, SDL_Rect rect2 );
     // lay ngau nhien trong x->y
